[Design] Idea for #6882
This design concept splits up UseBlazor into a usage of static files and
a usage of MapFallback.

---

I ignored auto-rebuild support because it's currently dead code until we
have that in VS. If we add auto-rebuild to ASP.NET Core - we'd probably want
to make that a separate gesture inside `IsDevelopement()` like other
dev-time features anyway.

---

The static-files wireup is pretty thick, and comes with some caveats.
This requires you to use the StaticFiles middleware without passing in
an explicit options instance. This gives us parity with what we're doing
for server-side, but it still gives me a bit of pause. Note that this
supports multiple calls to include multiple projects.

Another approach would be to make a separate `UseClientAppFiles()`
middleware. This wouldn't be quite so fragile.

Basically the approach in the PR optimizes the Startup code being
simpler but requiring a little more explanation. We could optimize the
other way by making an explicit middleware.

---

I feel pretty good about the wireup with routing to use the
`index.html` from the client app. I think it's pretty to-the-point.